Restructure helper files
