+++ e2e/tests/ui-driven/src/helpers/addComponent.ts
@@ -1,37 +1,4 @@
-import { Browser, Locator, Page, Request } from "@playwright/test";
-import { createAuthenticatedSession } from "../globalHelpers";
-
-export const isGetUserRequest = (req: Request) =>
-  req.url().includes("/user/me");
-
-export async function getAdminPage({
-  browser,
-  userId,
-}: {
-  browser: Browser;
-  userId: number;
-}): Promise<Page> {
-  const page = await createAuthenticatedSession({ browser, userId });
-  await page.goto("/");
-  await page.waitForResponse((response) => {
-    return response.url().includes("/graphql");
-  });
-  return page;
-}
-
-export async function getTeamPage({
-  browser,
-  userId,
-  teamName,
-}: {
-  browser: Browser;
-  userId: number;
-  teamName: string;
-}): Promise<Page> {
-  const page = await getAdminPage({ browser, userId });
-  await page.locator("h3", { hasText: teamName }).click();
-  return page;
-}
+import { Locator, Page } from "@playwright/test";
 
 export const createQuestionWithOptions = async (
   page: Page,
@@ -168,4 +135,3 @@ export const createContactInput = async (
     .filter({ hasText: "Create contact-input" })
     .click();
 };
-
