+++ e2e/tests/ui-driven/src/helpers/context.ts
@@ -1,8 +1,8 @@
-import assert from "node:assert";
-import { log } from "./globalHelpers";
-import { sign } from "jsonwebtoken";
 import { CoreDomainClient } from "@opensystemslab/planx-core";
 import { GraphQLClient, gql } from "graphql-request";
+import { sign } from "jsonwebtoken";
+import assert from "node:assert";
+import { log } from "./globalHelpers";
 
 type NewTeam = Parameters<CoreDomainClient["team"]["create"]>[0];
 
@@ -47,7 +47,7 @@ export const contextDefaults: Context = {
 };
 
 export async function setUpTestContext(
-  initialContext: Context,
+  initialContext: Context
 ): Promise<Context> {
   const $admin = getCoreDomainClient();
   const context: Context = { ...initialContext };
@@ -117,7 +117,7 @@ export function generateAuthenticationToken(userId: string) {
         "x-hasura-user-id": `${userId}`,
       },
     },
-    process.env.JWT_SECRET,
+    process.env.JWT_SECRET
   );
 }
 
@@ -127,7 +127,7 @@ export function getCoreDomainClient(): CoreDomainClient {
 
   const API = process.env.HASURA_GRAPHQL_URL!.replace(
     "${HASURA_PROXY_PORT}",
-    process.env.HASURA_PROXY_PORT!,
+    process.env.HASURA_PROXY_PORT!
   );
   const SECRET = process.env.HASURA_GRAPHQL_ADMIN_SECRET!;
   return new CoreDomainClient({
@@ -142,7 +142,7 @@ export function getGraphQLClient(): GraphQLClient {
 
 export async function findSessionId(
   adminGQLClient: GraphQLClient,
-  context,
+  context
 ): Promise<string | undefined> {
   // get the flow id which may have a session
   const flowResponse: { flows: { id: string }[] } =
@@ -152,7 +152,7 @@ export async function findSessionId(
           id
         }
       }`,
-      { slug: context.flow?.slug },
+      { slug: context.flow?.slug }
     );
   if (!flowResponse.flows.length || !flowResponse.flows[0].id) {
     return;
@@ -172,7 +172,7 @@ export async function findSessionId(
           }
         }
       `,
-      { flowId, email: context.user?.email },
+      { flowId, email: context.user?.email }
     );
   if (response.lowcal_sessions.length && response.lowcal_sessions[0].id) {
     return response.lowcal_sessions[0].id;
@@ -188,7 +188,7 @@ async function deleteSession(adminGQLClient: GraphQLClient, context) {
             id
           }
         }`,
-        { sessionId },
+        { sessionId }
       );
     }
   }
@@ -201,14 +201,14 @@ async function deleteSession(adminGQLClient: GraphQLClient, context) {
           id
         }
       }`,
-      { sessionId },
+      { sessionId }
     );
   }
 }
 
 async function deletePublishedFlow(
   adminGQLClient: GraphQLClient,
-  context: Context,
+  context: Context
 ) {
   if (context.flow?.publishedId) {
     log(`deleting published flow ${context.flow?.publishedId}`);
@@ -218,7 +218,7 @@ async function deletePublishedFlow(
           id
         }
       }`,
-      { publishedFlowId: context.flow?.publishedId },
+      { publishedFlowId: context.flow?.publishedId }
     );
   }
 }
@@ -232,7 +232,7 @@ async function deleteFlow(adminGQLClient: GraphQLClient, context: Context) {
           id
         }
       }`,
-      { flowId: context.flow?.id },
+      { flowId: context.flow?.id }
     );
   } else if (context.flow?.slug) {
     // try deleting via slug (when cleaning up from a previously failed test)
@@ -242,11 +242,11 @@ async function deleteFlow(adminGQLClient: GraphQLClient, context: Context) {
             id
           }
         }`,
-      { slug: context.flow?.slug },
+      { slug: context.flow?.slug }
     );
     if (response.flows.length && response.flows[0].id) {
       log(
-        `deleting flow ${context.flow?.slug} flowId: ${response.flows[0].id}`,
+        `deleting flow ${context.flow?.slug} flowId: ${response.flows[0].id}`
       );
       await adminGQLClient.request(
         `mutation DeleteTestFlow( $flowId: uuid!) {
@@ -254,7 +254,7 @@ async function deleteFlow(adminGQLClient: GraphQLClient, context: Context) {
             id
           }
         }`,
-        { flowId: response.flows[0].id },
+        { flowId: response.flows[0].id }
       );
     }
   }
@@ -269,7 +269,7 @@ async function deleteUser(adminGQLClient: GraphQLClient, context: Context) {
           id
         }
       }`,
-      { userId: context.user?.id },
+      { userId: context.user?.id }
     );
   } else if (context.user?.email) {
     // try deleting via email (when cleaning up from a previously failed test)
@@ -279,11 +279,11 @@ async function deleteUser(adminGQLClient: GraphQLClient, context: Context) {
           id
         }
       }`,
-      { email: context.user?.email },
+      { email: context.user?.email }
     );
     if (response.users.length && response.users[0].id) {
       log(
-        `deleting user ${context.user?.email} userId: ${response.users[0].id}`,
+        `deleting user ${context.user?.email} userId: ${response.users[0].id}`
       );
       await adminGQLClient.request(
         `mutation DeleteTestUser($userId: Int!) {
@@ -291,7 +291,7 @@ async function deleteUser(adminGQLClient: GraphQLClient, context: Context) {
             id
           }
         }`,
-        { userId: response.users[0].id },
+        { userId: response.users[0].id }
       );
     }
   }
@@ -306,7 +306,7 @@ async function deleteTeam(adminGQLClient: GraphQLClient, context: Context) {
           id
         }
       }`,
-      { teamId: context.team?.id },
+      { teamId: context.team?.id }
     );
   } else if (context.team?.slug) {
     // try deleting via slug (when cleaning up from a previously failed test)
@@ -316,11 +316,11 @@ async function deleteTeam(adminGQLClient: GraphQLClient, context: Context) {
                id
              }
            }`,
-      { slug: context.team?.slug },
+      { slug: context.team?.slug }
     );
     if (response.teams.length && response.teams[0].id) {
       log(
-        `deleting team ${context.team?.slug} teamId: ${response.teams[0].id}`,
+        `deleting team ${context.team?.slug} teamId: ${response.teams[0].id}`
       );
       await adminGQLClient.request(
         `mutation DeleteTestTeam( $teamId: Int!) {
@@ -328,7 +328,7 @@ async function deleteTeam(adminGQLClient: GraphQLClient, context: Context) {
           id
         }
       }`,
-        { teamId: response.teams[0].id },
+        { teamId: response.teams[0].id }
       );
     }
   }
@@ -353,7 +353,7 @@ async function setupGovPaySecret($admin: CoreDomainClient, context: Context) {
       {
         team_id: context.team.id,
         staging_govpay_secret: process.env.GOV_UK_PAY_SECRET_E2E,
-      },
+      }
     );
   } catch (error) {
     throw Error("Failed to setup GovPay secret for E2E team");
