+++ e2e/tests/ui-driven/src/helpers/getPage.ts
@@ -1,0 +1,31 @@
+import { Browser, Page } from "@playwright/test";
+import { createAuthenticatedSession } from "./globalHelpers";
+
+export async function getAdminPage({
+  browser,
+  userId,
+}: {
+  browser: Browser;
+  userId: number;
+}): Promise<Page> {
+  const page = await createAuthenticatedSession({ browser, userId });
+  await page.goto("/");
+  await page.waitForResponse((response) => {
+    return response.url().includes("/graphql");
+  });
+  return page;
+}
+
+export async function getTeamPage({
+  browser,
+  userId,
+  teamName,
+}: {
+  browser: Browser;
+  userId: number;
+  teamName: string;
+}): Promise<Page> {
+  const page = await getAdminPage({ browser, userId });
+  await page.locator("h3", { hasText: teamName }).click();
+  return page;
+}
